refactor(warnings): split reading order detectors
Move reading-order warning rules into a dedicated warnings domain module so the main warning catalog is easier to review by responsibility. The change keeps warning messages and detector order stable while reducing the size of the central warnings file.
